% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.ITime}
\alias{fread}
\alias{fwrite}
\alias{str_interp}
\alias{str_c}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table:IDateTime]{as.ITime}}, \code{\link[data.table]{fread}}, \code{\link[data.table]{fwrite}}}

  \item{stringr}{\code{\link[stringr]{str_c}}, \code{\link[stringr]{str_interp}}}
}}

