% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{SAILDB.METADATA}
\alias{SAILDB.METADATA}
\title{Metadata}
\format{
An object of class \code{list} of length 87.
}
\usage{
SAILDB.METADATA
}
\description{
A list of datasets, table references and other associated metadata to be used
as a \code{DatasetContainer} lookup & constants. Please use the \code{View()} function
to see the available options, \emph{e.g.} \code{View(SAILDB.METADATA)}
}
\details{
Each of the datasets are defined such that:
\preformatted{
# The key, i.e. [some.name], is the reserved dataset name
some.name   = list(
  ref    = 'some.name',  # an internal reference to the reserved name
  name   = 'SOME_NAME',  # the name of the table which may or may not contain a prefix, e.g. 'WLGP_'
  origin = 'SAILXXXXV',  # a reference to the base table from which this table can be found
  tag    = '_${date}',   # how refresh dates are tagged, i.e. some interpolated suffix; if \code{FALSE | NA} then no interpolation is performed
  alt    = 'WDSD',       # a prefix to the table name when defined as a character string; if \code{FALSE | NA} then no prefix will be added
  static = FALSE         # the \code{static} property describes whether the table is held by a reference schema instead of being copied to a project schema
)
}
}
\keyword{datasets}
