% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{Profile}
\alias{Profile}
\title{Profile}
\description{
User profile secret manager for SAIL DB authentication
}
\details{
Profile secret manager for SAIL DB authentication; exported to allow for use in your
personal projects and/or to better manage your own secrets
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{keychain}}{(\code{character|NA})\cr
A private reference to the keychain name}

\item{\code{connected}}{(\code{character|NA})\cr
A read-only field describing the name of the keychain}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{keychain}}{(\code{character|NA})\cr
A private reference to the keychain name}

\item{\code{connected}}{(\code{character|NA})\cr
A read-only field describing the name of the keychain}

\item{\code{system.user}}{(\code{character|NA})\cr
A read-only helper field to derive the client's system username}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Profile-new}{\code{Profile$new()}}
\item \href{#method-Profile-has.secrets}{\code{Profile$has.secrets()}}
\item \href{#method-Profile-get.secrets}{\code{Profile$get.secrets()}}
\item \href{#method-Profile-is.secret}{\code{Profile$is.secret()}}
\item \href{#method-Profile-set.secrets}{\code{Profile$set.secrets()}}
\item \href{#method-Profile-remove.secrets}{\code{Profile$remove.secrets()}}
\item \href{#method-Profile-clone}{\code{Profile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Profile-new"></a>}}
\if{latex}{\out{\hypertarget{method-Profile-new}{}}}
\subsection{Method \code{new()}}{
Initialise a user profile instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Profile$new(keychain.name = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keychain.name}}{(\code{character|NA})\cr
A keychain name; defaults to \code{SAILDB.DEF$KEYCHAIN} if no name is given}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new user profile instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Profile-has.secrets"></a>}}
\if{latex}{\out{\hypertarget{method-Profile-has.secrets}{}}}
\subsection{Method \code{has.secrets()}}{
Method to check whether a secret exists for the given username and database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Profile$has.secrets(username = NA, database = SAILDB.DEF$DATABASE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{(\code{character|NA})\cr
Optional username that relates to these secrets}

\item{\code{database}}{(\code{character|NA})\cr
The name of the database that these secrets relate to; defaults to \code{SAILDB.DEF$DATABASE} if no name is given}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical describing whether any secrets exist
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Profile-get.secrets"></a>}}
\if{latex}{\out{\hypertarget{method-Profile-get.secrets}{}}}
\subsection{Method \code{get.secrets()}}{
Method to collect the secrets associated with the given username and database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Profile$get.secrets(username = NA, database = SAILDB.DEF$DATABASE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{(\code{character|NA})\cr
A username that relates to these secrets}

\item{\code{database}}{(\code{character|NA})\cr
The name of the database that these secrets relate to; defaults to \code{SAILDB.DEF$DATABASE} if no name is given}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The secrets associated with this username & database if the key exists
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Profile-is.secret"></a>}}
\if{latex}{\out{\hypertarget{method-Profile-is.secret}{}}}
\subsection{Method \code{is.secret()}}{
Compare a known secret to the value stored in the keychain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Profile$is.secret(username = NA, password = NA, database = SAILDB.DEF$DATABASE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{(\code{character|NA})\cr
Optional username that relates to these secrets}

\item{\code{password}}{(\code{character|NA})\cr
The secret you would like to compare to any stored in the keychain}

\item{\code{database}}{(\code{character|NA})\cr
The name of the database that these secrets relate to; defaults to \code{SAILDB.DEF$DATABASE} if no name is given}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical describing the equivalence between the secrets
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Profile-set.secrets"></a>}}
\if{latex}{\out{\hypertarget{method-Profile-set.secrets}{}}}
\subsection{Method \code{set.secrets()}}{
Method to set the secrets for the given username and database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Profile$set.secrets(
  username = NA,
  password = NA,
  database = SAILDB.DEF$DATABASE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{(\code{character|NA})\cr
A username that relates to these secrets}

\item{\code{password}}{(\code{character|NA})\cr
The secret to store in the keychain}

\item{\code{database}}{(\code{character|NA})\cr
The name of the database that these secrets relate to; defaults to \code{SAILDB.DEF$DATABASE} if no name is given}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical describing whether this action was successful
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Profile-remove.secrets"></a>}}
\if{latex}{\out{\hypertarget{method-Profile-remove.secrets}{}}}
\subsection{Method \code{remove.secrets()}}{
Method to remove secrets associated with either (a) a username and a database; or (b) all secrets associated with a database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Profile$remove.secrets(username = NA, database = SAILDB.DEF$DATABASE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{(\code{character|NA})\cr
An optional username that relates to these secrets; if none are provided all of the secrets associated with the database will be removed}

\item{\code{database}}{(\code{character|NA})\cr
The name of the database that these secrets relate to; defaults to \code{SAILDB.DEF$DATABASE} if no name is given}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical describing whether this was successful
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Profile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Profile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Profile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
